% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatar_mes.R
\name{formatar_mes}
\alias{formatar_mes}
\title{Formatar mês}
\usage{
formatar_mes(data, label = TRUE)
}
\arguments{
\item{data}{Uma data ou uma string com uma data válida.}

\item{label}{Valor lógico. Se TRUE, retornará o nome do mês. Se FALSE,
o número.}
}
\value{
Se label = TRUE, retornará uma string. Se FALSE,
um número.
}
\description{
Formata uma data em um mês como número ou nome do mês.
}
